Add missing WMI query execution calls and fix MariaDB config path formatting
- Execute WMI queries in serviceExists() and getServiceState() methods that were missing ExecQuery() calls
- Remove unnecessary type cast from startup time minutes calculation as floor() already returns int-compatible value
- Fix MariaDB service configuration path by applying Windows path formatting to defaults-file parameter

+++ core/classes/actions/class.action.startup.php
@@ -216,7 +216,7 @@ class ActionStartup
             // Log total startup time in VERBOSE_TRACE mode (mode 3)
             global $bearsamppConfig;
             if ($bearsamppConfig->getLogsVerbose() == Config::VERBOSE_TRACE) {
-                $minutes = (int)floor($startupTime / 60);
+                $minutes = floor($startupTime / 60);  // floor() returns int-compatible value
                 $seconds = round($startupTime % 60, 2);
                 $formattedTime = sprintf('%d:%05.2f', $minutes, $seconds);
                 Util::logTrace('=== TOTAL STARTUP TIME: ' . $formattedTime . ' ===');
